new step renamed to a more suitable name (#24)
